Namespace keys and modify causes
One of the goal of this library is to ease the pain of sharing
exceptions over the network. Keywords in description maps are now
namespaced in order to be more rigorous. Theses keywords are specified
using clojure.spec so that it is obvious at first sight what information
a key refers to.

Describing an exception with `dvlopt.ex/exception` does not provide its
causes at :via anymore. Rather, `dlopt.ex/exception-and-causes`, returns
a lazy seq containing the exception and its causes starting with the
most direct one.